+++ tests/testthat/test-block.R
@@ -1,0 +1,11 @@
+test_that("empty description block is silently removed", {
+  expect_warning(
+    roc_proc_text(rd_roclet(), "
+      #'
+      #'
+      f <- function() {}
+      "
+    ),
+    NA
+  )
+})